Added images and changed way it is stored
Since maximum cursor size is limited and Base64 is so resource expensive (up to 33% more size than the file itself) I've decided to chunk  the images and store it in another table.
App is, without surprise, occupying double images (one copy in resources, one in database), but it has to be like that since if you wipe the app data, there's no backup to retrieve the images from.